+++ src/components/POLine/hooks/index.js
@@ -1,3 +1,4 @@
+export { useExchangeCalculation } from './useExchangeCalculation';
 export { useFundsById } from './useFundsById';
 export { useIsFundsRestrictedByLocationIds } from './useIsFundsRestrictedByLocationIds';
 export { useManageDonorOrganizationIds } from './useManageDonorOrganizationIds';